Move loadPlainTextLog.js into .../src/util
