+++ src/util/loadPlainTextLog.js
